refactor: use go:embed for storing crud templates
